requirements.txt
numpy
Pillow==9.2.0
streamlit==1.16.0
tensorflow==2.7.3